+++ RELEASE_NOTES.md
@@ -1,6 +1,10 @@
 Release notes
 ==============
 
+### vNext
+
+* `saveRoot` now takes insert/update functions that return `Async<unit>`; use the new `saveRootWithOutput` to return a value from insert/update
+
 ### 0.1.0 (2021-03-15)
 
 * Initial release